add doc and rename checkAuth to checkToken
