+++ src/api/auth/check-token.js
